For firetv call update user with device details upon initialization
