+++ iterableapi/src/main/java/com/iterable/iterableapi/IterableActivityMonitor.java
@@ -3,14 +3,13 @@ package com.iterable.iterableapi;
 import android.app.Activity;
 import android.app.Application;
 import android.content.Context;
-import android.os.Build;
 import android.os.Bundle;
 import android.os.Handler;
 import android.os.Looper;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
-import com.iterable.iterableapi.util.BrandUtils;
+import com.iterable.iterableapi.util.DeviceInfoUtils;
 
 import java.lang.ref.WeakReference;
 import java.util.ArrayList;
@@ -60,7 +59,7 @@ public class IterableActivityMonitor {
         public void onActivityResumed(Activity activity) {
             currentActivity = new WeakReference<>(activity);
 
-            if (!inForeground || BrandUtils.isFireTV(activity.getPackageManager())) {
+            if (!inForeground || DeviceInfoUtils.isFireTV(activity.getPackageManager())) {
                 inForeground = true;
                 for (WeakReference<AppStateCallback> callback : callbacks) {
                     if (callback.get() != null) {
