+++ iterableapi/src/main/java/com/iterable/iterableapi/IterableApi.java
@@ -11,6 +11,8 @@ import androidx.annotation.Nullable;
 import androidx.annotation.RestrictTo;
 import androidx.annotation.VisibleForTesting;
 
+import com.iterable.iterableapi.util.DeviceInfoUtils;
+
 import org.json.JSONException;
 import org.json.JSONObject;
 
@@ -543,6 +545,15 @@ public class IterableApi {
 
         loadLastSavedConfiguration(context);
         IterablePushNotificationUtil.processPendingAction(context);
+        if (DeviceInfoUtils.isFireTV(context.getPackageManager())) {
+            JSONObject dataFields = new JSONObject();
+            try {
+                DeviceInfoUtils.populateDeviceDetails(dataFields, context, sharedInstance.getDeviceId());
+                sharedInstance.apiClient.updateUser(dataFields, false);
+            } catch (JSONException e) {
+                    IterableLogger.e(TAG, "initialize: exception", e);
+            }
+        }
     }
 
     public static void setContext(Context context) {
