+++ iterableapi/src/main/java/com/iterable/iterableapi/IterableUtil.java
@@ -11,7 +11,7 @@ import org.json.JSONObject;
 
 import java.io.File;
 
-class IterableUtil {
+public class IterableUtil {
     @VisibleForTesting
     static IterableUtilImpl instance = new IterableUtilImpl();
 
@@ -19,11 +19,11 @@ class IterableUtil {
         return instance.currentTimeMillis();
     }
 
-    static String getAppVersion(Context context) {
+    public static String getAppVersion(Context context) {
         return instance.getAppVersion(context);
     }
 
-    static String getAppVersionCode(Context context) {
+    public static String getAppVersionCode(Context context) {
         return instance.getAppVersionCode(context);
     }
